cloud-function/requirements.txt
google-cloud-firestore
requests
openai
python-dotenv
